third_party/WebKit/LayoutTests/http/tests/devtools/elements/styles-3/style-autocomplete-expected.txt

Tests that autocompletions are computed correctly when editing the Styles pane.


Running: testEmptyName

Running: testEmptyNameForce

Running: testSingleCharName

Running: testSubstringName

Running: testEmptyValue

Running: testImportantDeclarationDoNotToggleOnExclamationMark

Running: testImportantDeclaration

Running: testValueR

Running: testValueWithParenthesis

Running: testValuePrefixed

Running: testValueUnprefixed

Running: testValueSubstring

Running: testNameVariables

Running: testValueVariables
